requirements.txt
EasyProcess==0.3
entrypoint2==0.2.3
MouseInfo==0.1.3
mss==6.1.0
Pillow==8.0.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
pyscreenshot==2.2
PyScreeze==0.1.26
PyTweening==1.0.3
six==1.15.0
xlib==0.21
